// apps/ui/airm/types/enums/project-workloads-table-field.ts
// Copyright © Advanced Micro Devices, Inc., or its affiliates.
//
// SPDX-License-Identifier: MIT

export enum ProjectWorkloadsTableField {
  NAME = 'displayName',
  TYPE = 'type',
  STATUS = 'status',
  GPUS = 'gpuCount',
  VRAM = 'vram',
  CREATED_AT = 'createdAt',
  CREATED_BY = 'createdBy',
  RUN_TIME = 'runTime',
}